+++ dokdo/api/__init__.py
@@ -15,7 +15,9 @@ from .beta_parallel_plot import beta_parallel_plot
 from .distance_matrix_plot import distance_matrix_plot
 from .taxa_abundance import taxa_abundance_bar_plot, taxa_abundance_box_plot
 from .ancom_volcano_plot import ancom_volcano_plot
-from .cross_association import cross_association_table, cross_association_heatmap, cross_association_regplot
+from .cross_association import (cross_association_table,
+    cross_association_heatmap, cross_association_regplot,
+    group_correlation_heatmap)
 
 from .addsig import addsig
 from .addpairs import addpairs
@@ -31,4 +33,5 @@ __all__ = ['alpha_diversity_plot', 'addpairs', 'wilcoxon',
            'taxa_abundance_bar_plot', 'taxa_abundance_box_plot',
            'ancom_volcano_plot', 'cross_association_table',
            'cross_association_heatmap', 'cross_association_regplot',
-           'addsig', 'regplot', 'addbiplot', 'ordinate', 'pname', 'get_mf']
+           'group_correlation_heatmap', 'addsig', 'regplot', 'addbiplot',
+           'ordinate', 'pname', 'get_mf']